// src/gen/coinbase/staking/v1alpha1/network.pb.ts
/* eslint-disable */
// @ts-nocheck
/*
* This file is a generated Typescript file for GRPC Gateway, DO NOT MODIFY
*/
export type Network = {
  name?: string
  isMainnet?: boolean
}

export type ListNetworksRequest = {
  parent?: string
}

export type ListNetworksResponse = {
  networks?: Network[]
}